move test of modelica version into seperate test file
